+++ src/render/deferred_renderer.cpp
@@ -64,7 +64,7 @@ void DeferredRenderer::set_gi_modifier_material_system(MaterialSystem* material_
         TextureInstance accumulator_texture;
         material_system->output(context(), 0, accumulator_texture);
         draw_material_system_->set_input(2, accumulator_texture);
-        material_system->set_priority(deferred_renderer_gbuffer_modifier_priority + priority);
+        material_system->set_priority(deferred_renderer_gbuffer_modifier_priority + static_cast<uint8_t>(priority));
     }
 }
